perf(web): load route surfaces on demand
The dashboard, projects, search and session detail were all static
imports, so every first paint downloaded and parsed markdown, syntax
highlighting and the 902-line receipt — even on an invalid route.

Each surface is now a dynamic import behind its own error boundary, the
receipt loads when its drawer opens, and the highlighter loads on the
first code output behind a fallback that already shows the code.
