+++ apps/web/src/components/tool-output/CodeHighlighter.tsx
@@ -1,80 +1,35 @@
-import SyntaxHighlighter from "react-syntax-highlighter/dist/esm/prism-light";
-import bash from "react-syntax-highlighter/dist/esm/languages/prism/bash";
-import c from "react-syntax-highlighter/dist/esm/languages/prism/c";
-import cpp from "react-syntax-highlighter/dist/esm/languages/prism/cpp";
-import csharp from "react-syntax-highlighter/dist/esm/languages/prism/csharp";
-import css from "react-syntax-highlighter/dist/esm/languages/prism/css";
-import docker from "react-syntax-highlighter/dist/esm/languages/prism/docker";
-import go from "react-syntax-highlighter/dist/esm/languages/prism/go";
-import ini from "react-syntax-highlighter/dist/esm/languages/prism/ini";
-import java from "react-syntax-highlighter/dist/esm/languages/prism/java";
-import javascript from "react-syntax-highlighter/dist/esm/languages/prism/javascript";
-import jsx from "react-syntax-highlighter/dist/esm/languages/prism/jsx";
-import json from "react-syntax-highlighter/dist/esm/languages/prism/json";
-import markdown from "react-syntax-highlighter/dist/esm/languages/prism/markdown";
-import markup from "react-syntax-highlighter/dist/esm/languages/prism/markup";
-import python from "react-syntax-highlighter/dist/esm/languages/prism/python";
-import rust from "react-syntax-highlighter/dist/esm/languages/prism/rust";
-import sql from "react-syntax-highlighter/dist/esm/languages/prism/sql";
-import toml from "react-syntax-highlighter/dist/esm/languages/prism/toml";
-import tsx from "react-syntax-highlighter/dist/esm/languages/prism/tsx";
-import typescript from "react-syntax-highlighter/dist/esm/languages/prism/typescript";
-import yaml from "react-syntax-highlighter/dist/esm/languages/prism/yaml";
-import oneDark from "react-syntax-highlighter/dist/esm/styles/prism/one-dark";
-import oneLight from "react-syntax-highlighter/dist/esm/styles/prism/one-light";
-import { useContext } from "react";
-import { ResolvedThemeContext } from "../../hooks/useTheme";
+import { lazy, Suspense } from "react";
+import { ErrorBoundary } from "../ErrorBoundary";
 
-SyntaxHighlighter.registerLanguage("bash", bash);
-SyntaxHighlighter.registerLanguage("c", c);
-SyntaxHighlighter.registerLanguage("cpp", cpp);
-SyntaxHighlighter.registerLanguage("csharp", csharp);
-SyntaxHighlighter.registerLanguage("css", css);
-SyntaxHighlighter.registerLanguage("docker", docker);
-SyntaxHighlighter.registerLanguage("go", go);
-SyntaxHighlighter.registerLanguage("ini", ini);
-SyntaxHighlighter.registerLanguage("java", java);
-SyntaxHighlighter.registerLanguage("javascript", javascript);
-SyntaxHighlighter.registerLanguage("jsx", jsx);
-SyntaxHighlighter.registerLanguage("json", json);
-SyntaxHighlighter.registerLanguage("markdown", markdown);
-SyntaxHighlighter.registerLanguage("markup", markup);
-SyntaxHighlighter.registerLanguage("python", python);
-SyntaxHighlighter.registerLanguage("rust", rust);
-SyntaxHighlighter.registerLanguage("sql", sql);
-SyntaxHighlighter.registerLanguage("toml", toml);
-SyntaxHighlighter.registerLanguage("tsx", tsx);
-SyntaxHighlighter.registerLanguage("typescript", typescript);
-SyntaxHighlighter.registerLanguage("yaml", yaml);
-
-interface CodeHighlighterProps {
+export interface CodeHighlighterProps {
   language: string;
   text: string;
 }
 
-function normalizeLanguage(language: string) {
-  if (language === "html" || language === "xml") return "markup";
-  return language;
+// Prism and its language grammars are the largest thing a session detail pulls
+// in, and plenty of sessions never show a code block — so it loads on the first
+// one, behind a fallback that already shows the code.
+const PrismHighlighter = lazy(() =>
+  import("./PrismHighlighter").then((m) => ({ default: m.PrismHighlighter })),
+);
+
+function PlainCode({ text }: { text: string }) {
+  return (
+    <pre
+      className="console-mono m-0 overflow-x-auto whitespace-pre-wrap break-words p-3 text-xs leading-relaxed"
+      style={{ background: "transparent" }}
+    >
+      {text}
+    </pre>
+  );
 }
 
 export function CodeHighlighter({ language, text }: CodeHighlighterProps) {
-  const resolvedTheme = useContext(ResolvedThemeContext);
   return (
-    <SyntaxHighlighter
-      language={normalizeLanguage(language)}
-      style={resolvedTheme === "dark" ? oneDark : oneLight}
-      customStyle={{
-        margin: 0,
-        padding: "0.75rem",
-        borderRadius: 0,
-        background: "transparent",
-        fontSize: "0.75rem",
-        lineHeight: 1.55,
-      }}
-      codeTagProps={{ className: "console-mono" }}
-      wrapLongLines
-    >
-      {text}
-    </SyntaxHighlighter>
+    <ErrorBoundary fallback={<PlainCode text={text} />}>
+      <Suspense fallback={<PlainCode text={text} />}>
+        <PrismHighlighter language={language} text={text} />
+      </Suspense>
+    </ErrorBoundary>
   );
 }
